refactor base text
